code reorg done but there are issues with linking to nested pages now as well as getting url params
