+++ v2/features/stats/jar/ChartContainer.tsx
@@ -1,13 +1,14 @@
 import React, { FC, useState } from "react";
-import { ChartSelect } from "./ChartSelect";
+import ChartSelect from "./ChartSelect";
 import LineChart from "./LineChart";
 import BiaxialChart from "./BiaxialChart";
 import YieldChart from "./MultiLineChart";
 import RevsChart from "./RevsChart";
-import { JarChartData } from "./types";
-import { TimeUnitPanel } from "./TimeUnitPanel";
+import TimeUnitPanel from "./TimeUnitPanel";
+import { JarChartData } from "v2/types";
 
-export const ChartContainer: FC<{ jarData: JarChartData }> = ({ jarData }) => {
+
+const ChartContainer: FC<{ jarData: JarChartData }> = ({ jarData }) => {
   const [selectedChart, setSelectedChart] = useState("value");
   const chartChange = (e: HTMLSelectElement): void => setSelectedChart(e.value);
 
@@ -55,3 +56,5 @@ export const ChartContainer: FC<{ jarData: JarChartData }> = ({ jarData }) => {
     </div>
   );
 };
+
+export default ChartContainer;
