+++ v2/features/stats/jar/TimeUnitPanel.tsx
@@ -1,5 +1,15 @@
 import React, { FC } from "react";
 
+const TimeUnitPanel: FC<{ timeChange: timeChangeFunction }> = ({ timeChange }) => {
+  return (
+    <span style={{ marginLeft: "80%" }}>
+      <TimeButton timeUnit="hr" displaytext="1h" timeChange={timeChange} />
+      <TimeButton timeUnit="hr6" displaytext="6h" timeChange={timeChange} />
+      <TimeButton timeUnit="day" displaytext="d" timeChange={timeChange} />
+      <TimeButton timeUnit="wk" displaytext="w" timeChange={timeChange} />
+    </span>
+  );
+};
 
 const TimeButton: FC<{
   timeUnit: string;
@@ -13,15 +23,4 @@ const TimeButton: FC<{
 
 type timeChangeFunction = (timeUnit: string) => void;
 
-export const TimeUnitPanel: FC<{ timeChange: timeChangeFunction }> = ({
-  timeChange,
-}) => {
-  return (
-    <span style={{ marginLeft: "80%" }}>
-      <TimeButton timeUnit="hr" displaytext="1h" timeChange={timeChange} />
-      <TimeButton timeUnit="hr6" displaytext="6h" timeChange={timeChange} />
-      <TimeButton timeUnit="day" displaytext="d" timeChange={timeChange} />
-      <TimeButton timeUnit="wk" displaytext="w" timeChange={timeChange} />
-    </span>
-  );
-};
+export default TimeUnitPanel;
